requirements.txt
praw==7.2.0

websocket-client==0.58.0
Reddit-ChatBot-Python==1.2.9
python-dotenv

